editing the modals for the bookings page
