+++ client/src/components/Calendar/EventDetail.jsx
@@ -1,22 +1,7 @@
 import { Dialog, Transition } from '@headlessui/react'
 import { Fragment, useState } from 'react'
 
-// components
-import DropDownMenu from './DropDownMenu'
-import RadioGroup from './RadioGroup'
-
-const SOURCE = [{ name: 'Cash', }, { name: 'Credit Card', }, { name: 'Voucher', }]
-const STATUS = [{ name: 'In Full', }, { name: 'Partial Payment', }, { name: 'No Payment', }]
-const RECEIVED = [{ name: 'No' }, { name: 'Yes' }]
-
 export default function EventDetail(props) {
-   const passengerNumbers = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-   const [adultNumber, setAdultNumber] = useState(0)
-   const [childrenNumber, setChildrenNumber] = useState(0)
-   const [infantNumber, setInfantNumber] = useState(0)
-   const [shirts, setShirts] = useState(0)
-   const [foodOptions, setFoodOption] = useState(0)
-
    let [isOpen, setIsOpen] = useState(false)
 
    function closeModal() {
@@ -64,6 +49,8 @@ export default function EventDetail(props) {
                         leaveTo="opacity-0 scale-95"
                      >
                         <Dialog.Panel className="w-full max-w-screen-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+                           
+                           {/* Date and time section */}
                            <div className='w-full h-auto rounded-[15px] bg-[#0E5BB5] p-4'>
                               <Dialog.Title
                                  as="h3"
@@ -77,51 +64,46 @@ export default function EventDetail(props) {
                               <h3 className='py-[10px] text-lg font-medium leading-6 text-gray-900'>
                                  When
                               </h3>
-                              <input type='date' className='border rounded-[10px] p-2'/>
+                              <input type='date' className='border rounded-[10px] p-2 w-1/4' />
 
                               <div className='inline-flex gap-4 items-center'>
-                                 <input type='time' className='border rounded-[10px] p-2'  />
+                                 <input type='time' className='border rounded-[10px] p-2 w-1/4' />
                                  <h1>to</h1>
-                                 <input type='time' className='border rounded-[10px] p-2' />
+                                 <input type='time' className='border rounded-[10px] p-2 w-1/4' />
                               </div>
                            </div>
 
+                           {/* Capacity section */}
                            <div className='py-[10px] flex flex-col'>
                               <h3 className='py-[10px] text-lg font-medium leading-6 text-gray-900'>
-                                 Contact Information *
+                                 Capacity
                               </h3>
-                              <div className='inline-flex gap-1'>
-                                 <input className='border rounded-[10px] p-2' placeholder='Email Address' />
-                                 <input className='border rounded-[10px] p-2' placeholder='Phone Number' />
-                              </div>
+                              <input className='border rounded-[10px] p-2 w-1/4' />
                            </div>
 
 
                            {/* Number of Passenger section*/}
                            <div className='py-[10px] flex flex-col'>
                               <h3 className='py-[10px] text-lg font-medium leading-6 text-gray-900'>
-                                 Passengers *
+                                 Passengers
                               </h3>
                               <div className='flex flex-col gap-2'>
-                                 <div className='inline-flex justify-between gap-1 items-center'>
-                                    <DropDownMenu list={passengerNumbers} setCurrent={setAdultNumber} current={adultNumber} />
-                                    <div>
+                                 <div className='inline-flex gap-5 items-center'>
+                                    <div className='w-1/12'>
                                        <h1 className='text-sm text-left font-medium leading-6 text-gray-900'>Adults</h1>
                                        <h1 className='text-xs text-left font-light text-gray-900'>Ages 12+</h1>
                                     </div>
                                     <input className='border rounded-[10px] p-2' placeholder='Price' />
                                  </div>
-                                 <div className='inline-flex justify-between gap-1 items-center'>
-                                    <DropDownMenu list={passengerNumbers} setCurrent={setChildrenNumber} current={childrenNumber} />
-                                    <div>
+                                 <div className='inline-flex gap-5 items-center'>
+                                    <div className='w-1/12'>
                                        <h1 className='text-sm text-left font-medium leading-6 text-gray-900'>Children</h1>
                                        <h1 className='text-xs text-left font-light text-gray-900'>Ages 5-11</h1>
                                     </div>
                                     <input className='border rounded-[10px] p-2' placeholder='Price' />
                                  </div>
-                                 <div className='inline-flex justify-between gap-1 items-center'>
-                                    <DropDownMenu list={passengerNumbers} setCurrent={setInfantNumber} current={infantNumber} />
-                                    <div>
+                                 <div className='inline-flex gap-5 items-center'>
+                                    <div className='w-1/12'>
                                        <h1 className='text-sm text-left font-medium leading-6 text-gray-900'>Infant</h1>
                                        <h1 className='text-xs text-left font-light text-gray-900'>Ages 0-4</h1>
                                     </div>
@@ -130,42 +112,13 @@ export default function EventDetail(props) {
                               </div>
                            </div>
 
-                           {/* Booking Addition section*/}
-                           <div className='py-[10px] flex flex-col'>
-                              <h3 className='py-[10px] text-lg font-medium leading-6 text-gray-900'>
-                                 Booking Additions
+                           {/* Event History */}
+                           <div className='py-[10px] flex flex-col overflow-scroll'>
+                              <h3 className='py-[10px] text-lg font-medium leading-6 text-gray-900 '>
+                                 Event History
                               </h3>
-                              <div className='inline-flex justify-between gap-2'>
-                                 <DropDownMenu list={passengerNumbers} setCurrent={setShirts} current={shirts} />
-                                 <div>
-                                    <h1 className='text-sm text-left font-medium leading-6 text-gray-900'>If you would like to purchase an EcoCat T-Shirt, select how many shirts you want. (Pick-up at check-in)</h1>
-                                    <h1 className='text-xs text-left font-light text-gray-900'>$19.99</h1>
-                                 </div>
-                              </div>
-                              <div className='inline-flex justify-between gap-2'>
-                                 <DropDownMenu list={passengerNumbers} setCurrent={setFoodOption} current={foodOptions} />
-                                 <div>
-                                    <h1 className='text-sm text-left font-medium leading-6 text-gray-900'>(FOOD OPTION) Select how many people want to upgrade to include our BBQ Hamburgers [NOTE: Base ticket does not include food. Please select this option if you wish to eat onboard!]</h1>
-                                    <h1 className='text-xs text-left font-light text-gray-900'>$9.99</h1>
-                                 </div>
-                              </div>
+                              
                            </div>
-                           <div className='inline-flex w-full justify-between'>
-                              <RadioGroup label={"Payment Source*"} plans={SOURCE} />
-                              <RadioGroup label={"Payment Status*"} plans={STATUS} />
-                              <RadioGroup label={"Commission Recieved*"} plans={RECEIVED} />
-                           </div>
-
-
-                           {/* Reservation Notes section*/}
-                           <div className='py-[10px] flex flex-col'>
-                              <h3 className='py-[10px] text-lg font-medium leading-6 text-gray-900'>
-                                 Reservation Notes
-                              </h3>
-                              <input className='border rounded-[10px] p-2 h-32' />
-                           </div>
-
-
 
                            <div className="mt-4">
                               <button
@@ -173,7 +126,7 @@ export default function EventDetail(props) {
                                  className="inline-flex justify-center rounded-md border border-transparent bg-blue-100 px-4 py-2 text-sm font-medium text-blue-900 hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2"
                                  onClick={closeModal}
                               >
-                                 CheckIn
+                                 Edit
                               </button>
                            </div>
                         </Dialog.Panel>
